Report rules on logical assets, if available
We want the sensor alerts to trigger on their logical assets. For T&H
sensors, this just works thanks to fty-metric-compute, which provides
the averages for racks, rows, etc and fty-alert-engine triggers alerts
based on those averages. For GPI metrics, there is no such
transformation, so we need to do it in the alert actor.

Note that the solution is not elegant, as there can be multiple assets
per rule in theory and only a single logical asset name, but in
practice, the array has a single element (the array itself is kind of a
hack in first place). Another minor issue is that the alert subject
still lists the sensorgpio asset, instead of the logical asset.